chore: delete tier 1 dead code from navigability audit (ENG-3724)
Removes ten dead-code items verified to have zero importers, per the
ENG-3684 fourth-pass navigability audit:

- A11 apps/journeys VideoWrapperPaused block wrapper
- A2 apps/journeys-admin findBlocksByTypename lib helper
- A4 apps/journeys-admin stray 0-byte src/__r file
- A7 apps/journeys-admin DiscoveryJourneys subtree (incl. EmbedJourney)
- A8 apps/journeys-admin ShareDrawer + LanguageScreenCardPreview
  (TemplateCustomization screens)
- A15 journeys-ui useSortLanguageContinents (dead twin of
  sortLanguageContinents)
- A16 journeys-ui Drawer/DrawerContent component
- A25 journeys-ui LanguageFilterDialog (live UI is
  LanguagesFilterPopper)
- A30 api-journeys getNextParentOrder (+ spec)
- A12 apps/journeys useFullscreenStatus hook, with its ambient
  declare global for vendor-prefixed fullscreen APIs relocated into
  EmbeddedPreview.tsx, the only consumer of those declarations

Type-check passes for journeys, journeys-admin, journeys-ui, and
api-journeys; lint passes for journeys.
