--- apps/journeys-admin/src/components/JourneyList/ActiveJourneyList/DiscoveryJourneys/DiscoveryJourneys.stories.tsx
@@ -1,21 +1,0 @@
-import { Meta, StoryObj } from '@storybook/nextjs-vite'
-
-import { journeysAdminConfig } from '@core/shared/ui/storybook'
-
-import { DiscoveryJourneys } from './DiscoveryJourneys'
-
-const DiscoveryJourneysStory: Meta<typeof DiscoveryJourneys> = {
-  ...journeysAdminConfig,
-  component: DiscoveryJourneys,
-  title: 'Journeys-Admin/JourneyList/DiscoveryJourneys'
-}
-
-const Template: StoryObj<typeof DiscoveryJourneys> = {
-  render: () => <DiscoveryJourneys />
-}
-
-export const Default = {
-  ...Template
-}
-
-export default DiscoveryJourneysStory
