--- apps/journeys-admin/src/components/JourneyList/ActiveJourneyList/DiscoveryJourneys/EmbedJourney/index.ts
@@ -1,1 +1,0 @@
-export { EmbedJourney } from './EmbedJourney'
