--- apps/journeys-admin/src/components/JourneyList/ActiveJourneyList/DiscoveryJourneys/index.ts
@@ -1,1 +1,0 @@
-export { DiscoveryJourneys } from './DiscoveryJourneys'
