--- apps/journeys-admin/src/components/TemplateCustomization/MultiStepForm/Screens/DoneScreen/ShareDrawer/index.ts
@@ -1,1 +1,0 @@
-export { ShareDrawer } from './ShareDrawer'
