--- apps/journeys-admin/src/components/TemplateCustomization/MultiStepForm/Screens/LanguageScreen/LanguageScreenCardPreview/index.ts
@@ -1,1 +1,0 @@
-export { LanguageScreenCardPreview } from './LanguageScreenCardPreview'
