--- apps/journeys-admin/src/libs/findBlocksByTypename/index.ts
@@ -1,1 +1,0 @@
-export { findBlocksByTypename } from './findBlocksByTypename'
