+++ apps/journeys/src/components/EmbeddedPreview/EmbeddedPreview.tsx
@@ -9,6 +9,23 @@ import { JourneyRenderer } from '../JourneyRenderer'
 
 import { ClickableCard } from './ClickableCard'
 
+declare global {
+  interface Document {
+    mozCancelFullScreen?: () => Promise<void>
+    msExitFullscreen?: () => Promise<void>
+    webkitExitFullscreen?: () => Promise<void>
+    mozFullScreenElement?: Element
+    msFullscreenElement?: Element
+    webkitFullscreenElement?: Element
+  }
+
+  interface HTMLElement {
+    msRequestFullscreen?: () => Promise<void>
+    mozRequestFullscreen?: () => Promise<void>
+    webkitRequestFullscreen?: () => Promise<void>
+  }
+}
+
 interface EmbeddedPreviewProps {
   disableFullscreen?: boolean
 }
