--- apps/journeys/src/components/VideoWrapperPaused/index.ts
@@ -1,1 +1,0 @@
-export { VideoWrapperPaused } from './VideoWrapperPaused'
