--- libs/journeys/ui/src/components/Drawer/index.ts
@@ -1,1 +1,0 @@
-export { Drawer, DrawerContent } from './Drawer'
