--- libs/journeys/ui/src/components/TemplateGallery/HeaderAndLanguageFilter/LanguageFilterDialog/index.ts
@@ -1,1 +1,0 @@
-export { LanguageFilterDialog } from './LanguageFilterDialog'
