--- libs/journeys/ui/src/libs/useSortLanguageContinents/index.ts
@@ -1,1 +1,0 @@
-export { useSortLanguageContinents } from './useSortLanguageContinents'
